Split UI bundle by route with lazy-loaded chunks
Cherry-picked from i386/codex/split-javascript-bundle-by-route with conflict
resolution. Adds remarkMath, mermaid/katex renderers, streaming prop to the
extracted chat-markdown component. Adds PR#11 row layout fix to the extracted
topology component. Adds remark-math to vite manualChunks config.
